restructured existing classes to follow MVVM architectural pattern for potential future additions
