+++ app/src/main/java/com/example/justeatcodingassignment/data/model/Restaurant.kt
@@ -1,4 +1,4 @@
-package com.example.justeatcodingassignment
+package com.example.justeatcodingassignment.data.model
 
 import com.squareup.moshi.Json
